+++ src/main/java/ro/ubbcluj/apm/threads/PrimeNumberCounterDemo.java
@@ -10,12 +10,13 @@ import java.util.concurrent.Future;
 public class PrimeNumberCounterDemo {
 
     public static final int N = 100_000_000;
-    public static final int NUMBER_OF_THREADS = 8;
+    public static final int NUMBER_OF_THREADS = 10;
     public static final int RANGE = N / NUMBER_OF_THREADS;
 
     public static void main(String[] args) {
         usingThreads();
         usingExecutorService();
+        usingVirtualThreads();
     }
 
     public static void usingExecutorService() {
@@ -34,8 +35,11 @@ public class PrimeNumberCounterDemo {
         for (Future<Integer> future : futures) {
             try {
                 totalPrimeCount += future.get();
-            } catch (InterruptedException | ExecutionException e) {
-                throw new RuntimeException(e);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                throw new RuntimeException("Thread was interrupted", e);
+            } catch (ExecutionException e) {
+                throw new RuntimeException("Exception during task execution", e);
             }
         }
 
@@ -59,6 +63,34 @@ public class PrimeNumberCounterDemo {
                 totalPrimeCount += thread.getPrimeCount();
             } catch (InterruptedException e) {
                 Thread.currentThread().interrupt();
+                throw new RuntimeException("Thread was interrupted", e);
+            }
+        }
+
+        System.out.println("Total prime numbers up to " + N + ": " + totalPrimeCount);
+    }
+
+    public static void usingVirtualThreads() {
+        List<Future<Integer>> futures = new ArrayList<>();
+
+        try (ExecutorService executorService = Executors.newVirtualThreadPerTaskExecutor()) {
+            for (int i = 0; i < NUMBER_OF_THREADS; i++) {
+                int start = i * RANGE;
+                int end = start + RANGE;
+                Future<Integer> future = executorService.submit(new PrimeCounterCallable(start, end));
+                futures.add(future);
+            }
+        }
+
+        int totalPrimeCount = 0;
+        for (Future<Integer> future : futures) {
+            try {
+                totalPrimeCount += future.get();
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                throw new RuntimeException("Thread was interrupted", e);
+            } catch (ExecutionException e) {
+                throw new RuntimeException("Exception during task execution", e);
             }
         }