+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 All notable changes to this project will be documented in this file.
 
+### Seminar 11
+
+- [x] Use volatile
+- [x] Compare with virtual threads
+
 ### Seminar 10
 
 - [x] Compare threads created by extending the Thread class and threads created by ExecutorService.